Cache FT/NFT counts (#454)
* Caching due to some super long-running queries to get the values

closes: #448

+++ service/src/main/kotlin/io/provenance/explorer/domain/models/explorer/AssetModels.kt
@@ -34,7 +34,7 @@ data class AssetDetail(
 )
 
 data class TokenCounts(
-    val fungibleCount: Long,
+    val fungibleCount: Int,
     val nonFungibleCount: Int
 )
